ProbeController method has been moved to service.go
